feat 뱃지에만 고양이 소리
