+++ my-local-diary/src/components/stamp/stamp.vue
@@ -15,7 +15,6 @@
         :src="stamp"
         alt="스탬프"
         class="stamp"
-        @click="playRandomMeow"
       />
     </div>
   </div>
@@ -23,7 +22,6 @@
 
 <script setup>
 import { computed } from 'vue';
-import { ref } from 'vue';
 import { defineProps } from 'vue';
 
 const props = defineProps({
@@ -35,31 +33,13 @@ const props = defineProps({
 const generatedStamps = computed(() => {
   return Array(props.count).fill(props.stampImage);
 });
-
-const playRandomMeow = () => {
-  const catSounds = [
-    'https://prismatic-sunshine-3445cc.netlify.app/cats/cat1.mp3',
-    'https://prismatic-sunshine-3445cc.netlify.app/cats/cat2.wav',
-    'https://prismatic-sunshine-3445cc.netlify.app/cats/cat3.mp3',
-    'https://prismatic-sunshine-3445cc.netlify.app/cats/cat4.mp3',
-    'https://prismatic-sunshine-3445cc.netlify.app/cats/cat5.mp3',
-    'https://prismatic-sunshine-3445cc.netlify.app/cats/cat6.mp3',
-    'https://prismatic-sunshine-3445cc.netlify.app/cats/cat7.mp3',
-    'https://prismatic-sunshine-3445cc.netlify.app/cats/cat8.mp3',
-    'https://prismatic-sunshine-3445cc.netlify.app/cats/cat9.mp3',
-    'https://prismatic-sunshine-3445cc.netlify.app/cats/cat10.wav',
-  ];
-  const randomIndex = Math.floor(Math.random() * catSounds.length);
-  const audio = new Audio(catSounds[randomIndex]);
-  audio.play();
-};
 </script>
 
 <style scoped>
 .cat-stamp-container {
   background-color: #fff5f7;
   padding: 10px;
-  width: 566px; /* 🔒 5개 도장이 딱 들어가는 고정 크기 */
+  width: 566px;
 }
 
 .category-bar {
@@ -92,15 +72,12 @@ const playRandomMeow = () => {
   height: 90px;
   display: block;
   border: none;
-  border-radius: 0;
-  padding: 0;
-  margin: 0;
   background-color: transparent;
   cursor: pointer;
   transition: transform 0.3s ease;
 }
 
 .stamp:hover {
-  transform: scale(1.3); /* 🐾 확실히 커지는 hover 효과 */
+  transform: scale(1.3);
 }
 </style>
